+++ client/src/components/meetups/MeetupFilterOptions/MeetupFilterOptions.module.scss
@@ -1,0 +1,82 @@
+@use '../../../styles/variables.scss' as *;
+
+.filters {
+  padding: 1.5rem;
+  background: white;
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+}
+
+.filter {
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+}
+
+.label {
+  color: $color-grey-600;
+}
+
+.options,
+.option {
+  display: flex;
+  align-items: center;
+}
+
+.options {
+  gap: 0.8rem;
+  flex-wrap: wrap;
+}
+
+.participants {
+  & .option {
+    gap: 1.5rem;
+  }
+
+  & .regulators {
+    color: $color-grey-600;
+    display: flex;
+    align-items: center;
+    gap: 1.5rem;
+
+    & .regulator {
+      display: flex;
+      align-items: center;
+
+      & span {
+        margin-left: 1rem;
+      }
+    }
+  }
+
+  & .tilde {
+    font-size: 1.8rem;
+  }
+}
+
+.number {
+  border: none;
+  padding: 0.5rem 0.7rem;
+  text-align: center;
+  width: 3.5rem;
+}
+.number::-webkit-outer-spin-button,
+.number::-webkit-inner-spin-button {
+  -webkit-appearance: none;
+}
+.number[type='number'] {
+  -moz-appearance: textfield;
+}
+
+.participation-fee-container {
+  & .option {
+    gap: 0.5rem;
+
+    & label {
+      cursor: pointer;
+
+      color: $color-grey-800;
+    }
+  }
+}
